+++ src/main/java/com/amartha/billing/engine/service/BillingEngine.java
@@ -4,7 +4,7 @@ package com.amartha.billing.engine.service;
 import com.amartha.billing.engine.model.Loan;
 
 public interface BillingEngine {
-  double getOutstanding(Loan loan);
+  long getOutstanding(Loan loan);
   boolean isDelinquent(Loan loan, int currentWeek);
-  boolean makePayment(Loan loan, double amount,  int currentWeek);
+  boolean makePayment(Loan loan, long amount, int currentWeek);
 }
